+++ haloface/apps/isomorphic/src/app/(hydrogen)/page.tsx
@@ -1,9 +1,9 @@
 import FileDashboard from '@/app/shared/file/dashboard';
-import { metaObject } from '@/config/site.config';
+// import { metaObject } from '@/config/site.config';
 
-export const metadata = {
-  ...metaObject(),
-};
+// export const metadata = {
+//   ...metaObject(),
+// };
 
 export default function FileDashboardPage() {
   // return <>Hello</>;
