+++ haloface/apps/isomorphic/src/app/auth/(forgot-password)/forgot-password-2/forgot-password-form.tsx
@@ -69,7 +69,7 @@ export default function ForgetPasswordForm() {
       <Text className="mt-5 text-center text-[15px] leading-loose text-gray-500 lg:text-start xl:mt-7 xl:text-base">
         Don’t want to reset?{' '}
         <Link
-          href={routes.auth.signIn2}
+          href={routes.auth.signIn}
           className="font-semibold text-gray-700 transition-colors hover:text-blue"
         >
           Sign In
