+++ haloface/apps/isomorphic/src/app/auth/(forgot-password)/forgot-password-2/page.tsx
@@ -1,4 +1,4 @@
-import AuthWrapperTwo from '@/app/shared/auth-layout/auth-wrapper-two';
+import AuthWrapperTwo from '@/app/shared/auth-layout/auth-wrapper';
 import ForgetPasswordForm from './forgot-password-form';
 
 export default function ForgotPassword() {
