+++ haloface/apps/isomorphic/src/app/auth/(sign-in)/sign-in/page.tsx
@@ -1,0 +1,14 @@
+import AuthWrapper from '@/app/shared/auth-layout/auth-wrapper';
+import SignInForm from './sign-in-form';
+//import { metaObject } from '@/config/site.config';
+// export const metadata = {
+//   ...metaObject('Sign In 2'),
+// };
+
+export default function SignIn() {
+  return (
+    <AuthWrapper title="Sign In" isSignIn isSocialLoginActive={true}>
+      <SignInForm />
+    </AuthWrapper>
+  );
+}
