+++ haloface/apps/isomorphic/src/app/auth/(sign-up)/sign-up/page.tsx
@@ -1,10 +1,10 @@
-import AuthWrapperTwo from '@/app/shared/auth-layout/auth-wrapper-two';
+import AuthWrapperTwo from '@/app/shared/auth-layout/auth-wrapper';
 import SignUpForm from './sign-up-form';
-import { metaObject } from '@/config/site.config';
+// import { metaObject } from '@/config/site.config';
 
-export const metadata = {
-  ...metaObject('Sign Up 2'),
-};
+// export const metadata = {
+//   ...metaObject('Sign Up 2'),
+// };
 
 export default function SignUpPage() {
   return (
