+++ haloface/apps/isomorphic/src/app/layout.tsx
@@ -38,7 +38,7 @@ export default async function RootLayout({
         suppressHydrationWarning
         className={cn(inter.variable, lexendDeca.variable, 'font-inter')}
       >
-        <AuthProvider session={session}>
+       <AuthProvider session={session}>
           <ThemeProvider>
             <NextProgress />
             <JotaiProvider>
