+++ haloface/apps/isomorphic/src/app/signin/page.tsx
@@ -1,16 +1,16 @@
 import SignInForm from '@/app/signin/sign-in-form';
-import AuthWrapperOne from '@/app/shared/auth-layout/auth-wrapper-one';
+import AuthWrapper from '@/app/shared/auth-layout/auth-wrapper';
 import Image from 'next/image';
 import UnderlineShape from '@core/components/shape/underline';
-import { metaObject } from '@/config/site.config';
+// import { metaObject } from '@/config/site.config';
 
-export const metadata = {
-  ...metaObject('Sign In'),
-};
+// export const metadata = {
+//   ...metaObject('Sign In'),
+// };
 
 export default function SignIn() {
   return (
-    <AuthWrapperOne
+    <AuthWrapper
       title={
         <>
           Welcome back! Please{' '}
@@ -27,22 +27,22 @@ export default function SignIn() {
       bannerDescription="Amet minim mollit non deserunt ullamco est sit aliqua dolor do
       amet sint velit officia consequat duis."
       isSocialLoginActive={true}
-      pageImage={
-        <div className="relative mx-auto aspect-[4/3.37] w-[500px] xl:w-[620px] 2xl:w-[820px]">
-          <Image
-            src={
-              'https://isomorphic-furyroad.s3.amazonaws.com/public/auth/sign-up.webp'
-            }
-            alt="Sign Up Thumbnail"
-            fill
-            priority
-            sizes="(max-width: 768px) 100vw"
-            className="object-cover"
-          />
-        </div>
-      }
+      // pageImage={
+      //   // <div className="relative mx-auto aspect-[4/3.37] w-[500px] xl:w-[620px] 2xl:w-[820px]">
+      //   //   <Image
+      //   //     src={
+      //   //       'https://isomorphic-furyroad.s3.amazonaws.com/public/auth/sign-up.webp'
+      //   //     }
+      //   //     alt="Sign Up Thumbnail"
+      //   //     fill
+      //   //     priority
+      //   //     sizes="(max-width: 768px) 100vw"
+      //   //     className="object-cover"
+      //   //   />
+      //   // </div>
+      // }
     >
       <SignInForm />
-    </AuthWrapperOne>
+    </AuthWrapper>
   );
 }
