+++ haloface/apps/isomorphic/src/config/site.config.tsx
@@ -19,26 +19,28 @@ export const siteConfig = {
   // TODO: favicon
 };
 
-export const metaObject = (
-  title?: string,
-  openGraph?: OpenGraph,
-  description: string = siteConfig.description
-): Metadata => {
-  return {
-    title: title ? `${title} - Isomorphic Furyroad` : siteConfig.title,
-    description,
-    openGraph: openGraph ?? {
-      title: title ? `${title} - Isomorphic Furyroad` : title,
-      description,
-      url: 'https://isomorphic-furyroad.vercel.app',
-      siteName: 'Isomorphic Furyroad', // https://developers.google.com/search/docs/appearance/site-names
-      images: {
-        url: 'https://s3.amazonaws.com/redqteam.com/isomorphic-furyroad/itemdep/isobanner.png',
-        width: 1200,
-        height: 630,
-      },
-      locale: 'en_US',
-      type: 'website',
-    },
-  };
-};
+// export const metaObject = (
+//   title?: string,
+//   openGraph?: OpenGraph,
+//   description: string = siteConfig.description
+// ): Metadata => {
+//   return {
+//     title: title ? `${title} - Isomorphic Furyroad` : siteConfig.title,
+//     description,
+//     openGraph: openGraph ?? {
+//       title: title ? `${title} - Isomorphic Furyroad` : title,
+//       description,
+//       url: 'https://isomorphic-furyroad.vercel.app',
+//       siteName: 'Isomorphic Furyroad', // https://developers.google.com/search/docs/appearance/site-names
+//       images: {
+//         url: 'https://s3.amazonaws.com/redqteam.com/isomorphic-furyroad/itemdep/isobanner.png',
+//         width: 1200,
+//         height: 630,
+//       },
+//       locale: 'en_US',
+//       type: 'website',
+//     },
+//   };
+// };
+
+
